requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
httpie==2.0.0
idna==2.9
psycopg2-binary==2.8.5
Pygments==2.6.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
